+++ common/src/main/java/net/mehvahdjukaar/polytone/item/ItemModelOverrideList.java
@@ -134,7 +134,7 @@ public class ItemModelOverrideList {
         }
 
     }
-
+/*
     public static void testTrie() {
         if (true) return;
         PropertiesSearchTrie trie = new PropertiesSearchTrie();
@@ -291,5 +291,5 @@ public class ItemModelOverrideList {
         System.out.println("this work " + search2);
 
     }
-
+*/
 }