+++ neoforge/src/main/java/net/mehvahdjukaar/polytone/neoforge/PlatStuffImpl.java
@@ -18,13 +18,7 @@ import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelManager;
 import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.client.renderer.block.model.ItemOverrides;
-import net.minecraft.client.resources.model.BakedModel;
-import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.core.MappedRegistry;
-import net.minecraft.core.Registry;
 import net.minecraft.core.RegistryAccess;
-import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.core.particles.ParticleType;
 import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.core.registries.BuiltInRegistries;
@@ -49,20 +43,6 @@ import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
 import net.neoforged.neoforge.common.CreativeModeTabRegistry;
 import net.neoforged.neoforge.common.world.ModifiableBiomeInfo;
 import net.neoforged.neoforge.server.ServerLifecycleHooks;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.DimensionSpecialEffectsManager;
-import net.minecraftforge.client.event.ModelEvent;
-import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
-import net.minecraftforge.common.CreativeModeTabRegistry;
-import net.minecraftforge.common.world.ModifiableBiomeInfo;
-import net.minecraftforge.fml.ModList;
-import net.minecraftforge.fml.ModLoader;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.loading.FMLEnvironment;
-import net.minecraftforge.fml.util.ObfuscationReflectionHelper;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.ForgeRegistry;
-import net.minecraftforge.server.ServerLifecycleHooks;
 import org.joml.Vector3f;
 
 import java.util.List;
@@ -272,32 +252,20 @@ public class PlatStuffImpl {
     }
 
 
-    public static RegistryAccess getServerRegistryAccess(){
+    public static RegistryAccess getServerRegistryAccess() {
         return ServerLifecycleHooks.getCurrentServer().registryAccess();
     }
 
-    public static BakedModel getBakedModel(ResourceLocation id) {
-        return Minecraft.getInstance().getModelManager().getModel(ModelResourceLocation.standalone(id));
+    public static BakedModel getBakedModel(ModelResourceLocation id) {
+        ModelManager mm = Minecraft.getInstance().getModelManager();
+        return mm.getModel(id);
     }
 
     public static void addSpecialModelRegistration(Consumer<PlatStuff.SpecialModelEvent> eventListener) {
         Consumer<ModelEvent.RegisterAdditional> eventConsumer = event -> {
-            eventListener.accept(new PlatStuff.SpecialModelEvent() {
-                @Override
-                public void register(ModelResourceLocation modelLocation) {
-                    event.register(modelLocation);
-                }
-
-                @Override
-                public void register(ResourceLocation id) {
-                    event.register((id));
-                }
-            });
+            eventListener.accept(event::register);
         };
         PolytoneForge.bus.addListener(eventConsumer);
     }
 
-    public static BakedModel getModel(ResourceLocation modelLocation) {
-        return Minecraft.getInstance().getModelManager().getModel(modelLocation);
-    }
 }